Advent

Description

Haskell bindings for Advent of Code 2018 REST API.

Specify your requests with AoC and AoCSettings, and run them with runAoC.

Please use responsibly. All actions are rate-limited to a minimum of one request per minute.

API

data AoC :: Type -> Type where

An API command. An AoC a an AoC API request that returns results of type a.

A lot of these commands take Finite 25, which represents a day of December up to and including Christmas Day (December 25th). You can convert an integer day (1 - 25) into a Finite 25 representing that day using mkDay or mkDay_.

Constructors

AoCPrompt
Fetch prompts, as HTML.

Note that you can give a bogus Session Key and still get Part 1 of a given day, provided the day is already released.

Fields

:: Finite 25

Day.

-> Char

Part. 'a' for part 1, 'b' for part 2, etc.

-> AoC Text

Prompt (as HTML)

AoCInput
Fetch input

Fields

:: Finite 25

Day.

-> AoC Text

AoCSubmit
Submit answer.

WARNING: Answers are not length-limited. Answers are stripped of leading and trailing whitespace and run through encode before submitting.

Fields

:: Finite 25

Day.

-> Char

Part. 'a' for part 1, 'b' for part 2, etc.

-> String

Answer. Is not length-limited.

-> AoC (Text, SubmitRes)

Submission reply (as HTML), and result token

data AoCSettings

Setings for running an API request.

Session keys are required for all commands, but if you enter a bogus key you should be able to get at least Part 1 from AoCPrompt.

The session key can be found by logging in on a web client and checking the cookies. You can usually check these with in-browser developer tools.

Throttling is hard-limited to a minimum of 1 second between calls. Please be respectful and do not try to bypass this.

If no cache directory is given, one will be allocated using getTemporaryDirectory.

Constructors

AoCSettings

Fields

_aSessionKey :: String

Session key

_aYear :: Integer

Year of challenge

_aCache :: Maybe FilePath

Cache directory

_aThrottle :: Int

Throttle delay, in milliseconds. Minimum is 1000000.

data SubmitRes

The result of a submission.

Constructors

SubCorrect (Maybe Integer)
global rank

SubIncorrect

SubWait

SubInvalid

SubUnknown

showSubmitRes :: SubmitRes -> String

Pretty-print a SubmitRes

runAoC :: AoCSettings -> AoC a -> IO (Either AoCError a)

Run an API command with a given AoCSettings to produce the result or a list of (lines of) errors.

WARNING: Answers are not length-limited. Answers are stripped of leading and trailing whitespace and run through encode before submitting.

defaultAoCSettings :: Integer -> String -> AoCSettings

Sensible defaults for AoCSettings for a given year and session key.

Calendar

challengeReleaseTime

Arguments

:: Integer
year

-> Finite 25
day

-> UTCTime

Prompt release time

timeToRelease

Arguments

:: Integer
year

-> Finite 25
day

-> IO NominalDiffTime

Get time until release of a given challenge.

challengeReleased

Arguments

:: Integer
year

-> Finite 25
day

-> IO Bool

Check if a challenge has been released yet.

Utility

Day

mkDay :: Integer -> Maybe (Finite 25)

Construct a Finite 25 (the type of a Day) from a day integer (1 - 25). If input is out of range, Nothing is returned. See mkDay_ for an unsafe version useful for literals.

Inverse of dayToInt.

mkDay_ :: Integer -> Finite 25

Construct a Finite 25 (the type of a Day) from a day integer (1 - 25). Is undefined if input is out of range. Can be useful for compile-time literals, like mkDay_ 4

Inverse of dayToInt.

aocDay :: AoC a -> Finite 25

Get the day associated with a given API command.

Throttler

setAoCThrottleLimit :: Int -> IO ()

Set the internal throttler maximum queue capacity. Default is 1000.

getAoCThrottleLimit :: IO Int

Get the internal throttler maximum queue capacity.

Internal

parseSubmitRes :: Text -> SubmitRes

Parse Text into a SubmitRes.